requirements.txt
# FastAPI 框架與ASGI伺服器
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic[email]>=2.4.2      # 加上 [email] 以啟用 email-validator
pydantic-settings>=2.0.3
email-validator>=1.3.1      # 明確列出；若只用 pydantic[email] 也會自動安裝

# 資料庫ORM與遷移
sqlalchemy>=2.0.21
alembic>=1.12.0
psycopg2-binary>=2.9.7
asyncpg>=0.28.0

# 認證與安全
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# 外部服務整合
aiohttp>=3.8.5
emails>=0.6.0
jinja2>=3.1.2
reportlab>=4.0.4
python-dotenv>=1.0.0
line-bot-sdk>=3.3.0

# 測試
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.25.0

# 開發工具
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
